python/requirements.txt
# Python OCR 의존성

# OCR 라이브러리 (예시 - 실제 클라이언트 제공 코드에 맞게 수정)
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# PDF 처리
PyPDF2==3.0.1

# JSON 처리 (기본 내장)

# ===================
# 공통 데이터 처리
# ===================
pandas==2.1.4
numpy==1.26.2

# ===================
# 유사물건 추천
# ===================
PyYAML==6.0.1

# Supabase 클라이언트 (추천 로직에서 DB 조회용)
supabase==2.3.4

